# docs/应用集成/OpenAPI/warehouse.openapi.yaml
openapi: 3.1.0
info:
  title: Warehouse HTTP API
  version: 1.0.0
  description: |
    Warehouse 对外 JSON HTTP API 的正式契约。

    WebDAV (`/dav/*`) 和独立 S3 Endpoint 是协议适配层，不在本 OpenAPI
    规范中逐操作重复定义。WebDAV 请参阅用户使用指南，S3 请参阅
    `docs/应用集成/S3协议.md`。
servers:
  - url: https://warehouse.tidukongjian.com
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/public/assets/spaces:
    get:
      tags: [Assets]
      operationId: listAssetSpaces
      summary: 获取当前用户的逻辑资产空间
      responses:
        "200":
          description: 资产空间列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetSpacesEnvelope"
        "401":
          $ref: "#/components/responses/SDKUnauthorized"
  /api/v1/public/assets/object:
    get:
      tags: [Assets]
      operationId: getAssetObjectMetadata
      summary: 获取当前用户资产对象元数据
      description: |
        通过 `/personal`、`/apps` 或 `/services` 开头的逻辑路径获取对象元数据。
        该接口用于 Knowledge 等上层服务读取 Warehouse 原始资产或回写资产的可验证元信息。
      parameters:
        - $ref: "#/components/parameters/AssetObjectPath"
      responses:
        "200":
          description: 对象元数据
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AssetObject"}
        "400": {$ref: "#/components/responses/AssetObjectError"}
        "401": {$ref: "#/components/responses/AssetObjectError"}
        "403": {$ref: "#/components/responses/AssetObjectError"}
        "404": {$ref: "#/components/responses/AssetObjectError"}
  /api/v1/public/assets/object/content:
    get:
      tags: [Assets]
      operationId: downloadAssetObjectContent
      summary: 下载当前用户资产对象内容
      parameters:
        - $ref: "#/components/parameters/AssetObjectPath"
      responses:
        "200":
          description: 对象内容
          headers:
            ETag:
              schema: {type: string}
            X-Warehouse-Checksum-SHA256:
              schema: {type: string}
            Content-Length:
              schema: {type: integer, format: int64}
          content:
            application/octet-stream:
              schema: {type: string, format: binary}
        "400": {$ref: "#/components/responses/AssetObjectError"}
        "401": {$ref: "#/components/responses/AssetObjectError"}
        "403": {$ref: "#/components/responses/AssetObjectError"}
        "404": {$ref: "#/components/responses/AssetObjectError"}
    head:
      tags: [Assets]
      operationId: headAssetObjectContent
      summary: 获取当前用户资产对象内容头信息
      parameters:
        - $ref: "#/components/parameters/AssetObjectPath"
      responses:
        "200":
          description: 对象头信息
          headers:
            ETag:
              schema: {type: string}
            X-Warehouse-Checksum-SHA256:
              schema: {type: string}
            Content-Length:
              schema: {type: integer, format: int64}
        "400": {$ref: "#/components/responses/AssetObjectError"}
        "401": {$ref: "#/components/responses/AssetObjectError"}
        "403": {$ref: "#/components/responses/AssetObjectError"}
        "404": {$ref: "#/components/responses/AssetObjectError"}
    put:
      tags: [Assets]
      operationId: putAssetObjectContent
      summary: 写入当前用户资产对象内容
      description: |
        写入必须落在 `/personal`、`/apps` 或 `/services` 下，并复用现有 quota、
        mutation recorder 和对象 metadata 逻辑。`X-Warehouse-Checksum-SHA256`
        可传 hex 或 base64 编码的 SHA-256；提供时必须校验通过。
      parameters:
        - $ref: "#/components/parameters/AssetObjectPath"
        - name: X-Warehouse-Checksum-SHA256
          in: header
          required: false
          schema: {type: string}
          description: hex 或 base64 编码的 SHA-256 校验值
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: {type: string, format: binary}
          text/plain:
            schema: {type: string}
          application/json:
            schema: {}
      responses:
        "200":
          description: 写入成功后的对象元数据
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AssetObject"}
        "400": {$ref: "#/components/responses/AssetObjectError"}
        "401": {$ref: "#/components/responses/AssetObjectError"}
        "403": {$ref: "#/components/responses/AssetObjectError"}
        "413": {$ref: "#/components/responses/AssetObjectError"}
  /api/v1/public/assets/objects:
    get:
      tags: [Assets]
      operationId: listAssetObjects
      summary: 列出当前用户资产对象
      parameters:
        - name: prefix
          in: query
          required: true
          schema:
            type: string
            pattern: "^/(personal|apps|services)(/.*)?$"
          example: /services/knowledge/
        - name: delimiter
          in: query
          required: false
          schema:
            type: string
            enum: ["/"]
          description: 传 `/` 时返回一层 common prefixes
      responses:
        "200":
          description: 对象列表
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AssetObjectList"}
        "400": {$ref: "#/components/responses/AssetObjectError"}
        "401": {$ref: "#/components/responses/AssetObjectError"}
        "403": {$ref: "#/components/responses/AssetObjectError"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or UCAN
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
...
  parameters:
    AssetObjectPath:
      name: path
      in: query
      required: true
      schema:
        type: string
        pattern: "^/(personal|apps|services)/.+"
      example: /services/knowledge/artifacts/report.md
  responses:
...
    SDKUnauthorized:
      description: 未认证或凭据无效
      content:
        application/json:
          schema: {$ref: "#/components/schemas/SDKError"}
...
    AssetObjectError:
      description: 资产对象接口错误
      content:
        application/json:
          schema: {$ref: "#/components/schemas/AssetObjectError"}
...
  schemas:
...
    SDKError:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          oneOf: [{type: integer}, {type: string}]
        message: {type: string}
        data: {type: [object, "null"], additionalProperties: true}
        timestamp: {type: integer, format: int64}
...
    SDKEnvelope:
      type: object
      required: [code, message, data, timestamp]
      properties:
        code: {type: integer, const: 0}
        message: {type: string, const: ok}
        data: {}
        timestamp: {type: integer, format: int64}
    AssetSpace:
      type: object
      required: [key, name, path]
      properties:
        key: {type: string, enum: [personal, apps, services]}
        name: {type: string}
        path: {type: string, pattern: "^/(personal|apps|services)$"}
    AssetSpacesEnvelope:
      allOf:
        - $ref: "#/components/schemas/SDKEnvelope"
        - type: object
          properties:
            data:
              type: object
              required: [defaultSpace, spaces]
              properties:
                defaultSpace: {type: string}
                spaces:
                  type: array
                  items: {$ref: "#/components/schemas/AssetSpace"}
    AssetObject:
      type: object
      required: [path, bucket, key, size, etag, contentType, modifiedAt, isPrefix]
      properties:
        path:
          type: string
          pattern: "^/(personal|apps|services)/.+"
        bucket:
          type: string
          enum: [personal, apps, services]
        key: {type: string}
        size: {type: integer, format: int64, minimum: 0}
        etag: {type: string}
        checksumSha256:
          type: string
          description: hex 编码的 SHA-256，仅单对象 metadata、download、write 响应返回
        contentType: {type: string}
        modifiedAt: {type: string, format: date-time}
        isPrefix: {type: boolean}
    AssetObjectList:
      type: object
      required: [prefix, objects, prefixes]
      properties:
        prefix:
          type: string
          pattern: "^/(personal|apps|services)(/.*)?$"
        objects:
          type: array
          items: {$ref: "#/components/schemas/AssetObject"}
        prefixes:
          type: array
          items:
            type: string
            pattern: "^/(personal|apps|services)/.*"
    AssetObjectError:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
...
security:
  - bearerAuth: []